Allow NCCL communicator creation to be canceled.
Recall that to perform NCCL collectives (e.g., AllReduce, AllGather), a process
creates a NCCL communicator and then uses the communicator to perform
collectives. If a participant of a collective fails, the collective can hang
forever. However, we can abort a NCCL communicator which aborts all pending
collectives being performed on it.

Creating a communicator can also block if a participant fails, but the code
didn't support canceling the creation of the communicator. This change allows
for communicator creation to be aborted.

The change isn't very deep. It's mostly just plumbing around some
`atomic_bool`s which are set to true when we want to cancel.

PiperOrigin-RevId: 808714413