Move custom call handler resolution into CustomCallThunk
For a custom call we need to resolve a target name (string) to a function pointer. So far this happens in `IrEmitterUnnested` (at the end of the compilation pipeline). But for thunk serialization we need this to happen at runtime (when the thunks are getting reconstructed from the proto representation). Therefore I'm moving this resolving step into the `CustomCallThunk` factory function.

Note that there remains a way to construct a `CustomCallThunk` from just a function pointer. These thunks will not be serializable and that's okay. The logic handles these cases and returns an error. It is important for tests to be able to quickly create a CustomCallThunk from a closure. If we had to register these calls in the registry first it would complicate our tests significantly.

In detail this change entails:
- Move resolver logic in new overloads of the `CustomCallThunk::Create` factory function.
- Call these overloads from `IrEmitterUnnested` and the custom kernel fusion emitter.
- Add tests for the new overloads
- Migrate some tests to FFI (the new custom call registry and API)
- Adjust some error codes in custom call tests. (If a custom call was not found now `kNotFound` is returned instead of `kUnimplemented`).

PiperOrigin-RevId: 814223143